Add GetVersion to GoogleKeymaster implementation.
The GetVersion command will be used by the GoogleKeymaster HAL
implementation to verify that HAL implementation and keymaster
implementation are compatible.  Note that the GetVersion API is not
exposed through the HAL, and there's no requirement that other keymaster
implementations provide versioning (it's a good idea, though).

This change also enables message versioning support.

Bug: 18409838
Change-Id: Ia22791356db3453deab4c21f06a16a08fac32d0e