Fix heap-smashing bug in keymaster key encryption.
If ae_init ever fails, the memset intended to clean up the OCB key
structure would actually zero a good chunk of random heap space because
zeros from the address of the pointer to the key structure, rather than
the address of the key structure.  This fixes it.

This change was already reviewed, merged and reverted, so I'm skipping
the review step this time.

Change-Id: If165f5583c588b695725f7c30f6d54d7805ba9dc