More fixes for BoringSSL compilation.
EC_GROUP_set_point_conversion_form has been removed in BoringSSL because
it didn't do anything.

Also, BoringSSL uses size_t and keymaster builds with a signed/unsigned
mismatch as a fatal error. This means that the casts to int aren't
needed in BoringSSL and, in fact, cause an error.

Change-Id: I52b7d34a5c90f40cfcc84c60b746404f374b1e80