Rename soft_keymaster_device to libsoftkeymaster
libsoftkeymaster will be loaded by keystore as a fallback device to use
when the hardware keymaster doesn't support certain types of keys.

This also moves the lib from /lib/hw to /lib since it doesn't currently
work for loading via HAL and /lib/hw isn't in the default library path.

Change-Id: Id32418eccde7997479e5a122c047c2f06bff6475