Change handling of debug log statments in non-debug builds.
The previous way had a problem when used in statments like:

if (foo)
   LOG_D(...);

When built without debugging, this became:

if (foo)
   ;

Which is sort of okay, but the compiler complains.  The new way also has
the advantage that the compiler always sees and checks the log
arguments.  Given that it ends up compiling something like:

do {
    if (0)
        Logger::Debug(...);
} while (0);

It should optimize the entire block out, and should even discard the
literal string used for the format.  So it's better all around.

Cherry-picked from internal.

Change-Id: I1a1cdd9481530ee3df8c96442cc1adf845b4c2c8