Don't bother parsing keymaster0 signing params.
The keymaster0 signing API includes a void* to a structure that depends
on key type (RSA or EC).  Previously we've tried to determine what the
key type is and to extract proper parameters, by calling
get_key_characteristics and examining the result.  But this is all
pointless.  The possible contents of that signing structure is fixed, so
we may as well just set the values directly.  This does mean that we set
KM_TAG_PADDING for EC keys, even though ECDSA doesn't use padding.
That's okay, though.  Keymaster1 implementations should ignore
extraneous tags.  And in any case, we'll soon stop even providing the
keymaster0 APIs, so the issue will disappear.

Change-Id: Id4ee756dedc716978323767d51b4e3db357da662