Add keymaster2 stub support to SoftKeymasterDevice.
This CL adds a keymaster2 API to SoftKeymasterDevice.  It just
delegates to the keymaster1 implementation, and will return errors if
any keymaster2 features are used. This will allow conversion of
keystore to the keymaster2 API, after which the keymaster1 API will be
removed from SoftKeymasterDevice (though it will still be able to wrap
a keymaster1 hardware implementation).

Note that this CL does not update the unit tests to test the
keymaster2 interface. That will come in a future CL.

Change-Id: I2efd696d733c99e916f98ae68586310d3b2abf78