Refactor keymaster tags.
This refactor separates declaration and definition of the tag
instances, so they don't get duplicated in every compilation unit.
Also, if tag names are enabled it provides a dynamically-generated map
from tags to names. This is only for testing/debugging use, but it's
very convenient for that. The test utils are updated to take advantage
of it.

Change-Id: I13ad2564e4a850c6b268a2cd1b3f5a37037ea6b7