system/keymaster: don't pass a structure into |d2i_PrivateKey|.
Some OpenSSL parsing functions have, historically, allowed a structure
to be passed in to reuse that memory. There have been many bugs arising
from this corner case and it's generally best to avoid it.

This change just passes in NULL because a new structure was being
allocated anyway. Also, the API didn't guarantee that the memory would
always be reused – code had to check the updated pointer, which this
didn't do. So it might have broken in the future.

(This change mirrors
https://android-review.googlesource.com/#/c/196590/1)

Change-Id: I5fed1020997cb8708706bbeb37a1e9bb1e6f71d3