Fix reboots due to unsafe key blob handling in legacy keymaster HAL
Some key blobs seem to upset certain keymaster implementations
such that they cause the device to reboot rather than reporting an
invalid key blob.

This patch removes the fast path to hardware from the software
keymaster implementation of get_key_characteristics, so that key blobs
are properly identified before passed to the correct engine.

This avoids potentially dangerous blobs being passed to keymaster
HALs for the purpose of retrieving key characteristics, that is.

Test: Fix for keystore CTS test
Change-Id: I3aa61d17c72f0319f9e1b682a8bdcf67416d5aea