Fix unsafe AuthorizationSet copy constructor
The copy constructor copies the internal error code after calling
Reinitialize which may fail. The implications are twofold.
A possible failure of Reinitialize may get masked.
Reinitialize may get called on unsafe data in the case where
the argument was invalid.

After this fix, the error state of the argument is copied first,
then checked, and iff OK, Reinitialize is called. A potential
failure of Reinitialize is exposed as expected.

Change-Id: I79d94109d9046ce101788e48c1ed108ec42d697f