Allow softkeymaster to use default AES key factory.
This is needed so that attempts to attest to an AES key fail correctly
when softkeymaster is wrapping a keymaster1 implementation.  The idea of
not having an AES factory at all to provoke errors if it's inadvertently
used was nice, but we have adequate testing in place to verify that
nothing like that happens without us noticing that AES keys are in
software when they should be in hardware (by checking key
characteristics).

Test: Tested by VTS and CTS
Change-Id: If3d340c3c462f559f5463129838a6673fe78286d