Don't check token for NO_AUTH_REQUIRED
IF NO_AUTH_REQUIRED or AUTH_TIMEOUT is specified, we don't need to check
the auth token. We began checking the auth token even in those cases
when the TRUSTED_CONFIRMATION_REQUIRED tag was introduced because the
TRUSTED_CONFIRMATION_REQUIRED tag must still be checked in those cases.

cherry-picked from:
https://android-review.googlesource.com/c/platform/system/keymaster/+/950602

Change-Id: I7b1fe55527f344759480330c2bb6f0a75ef15ced